requirements.txt
# For image manipulation and processing
Pillow

# For screenshot capture
mss

# For interacting with windows
pygetwindow

# For automating GUI interactions
pyautogui
pyHM

# For working with object detection models
python-opencv

# For working with URLs
requests

# Optional dependencies for the custom_logger if color support is needed
colorama

# deep learning dependencies
transformers
fastapi
uvicorn
python-multipart
